reorganize file struture
